refactor: 파일 구조 변경
atomic design 에서 한 폴더에는 하나의 컴포넌트(index.tsx)만 가지도록 구조 변경
css 약간의 수정(보이는건 차이 없음)
