+++ frontend/src/components/organisms/ChatContent/styles.ts
@@ -7,10 +7,11 @@ interface Props {
 export const Container = styled.div<Props>`
   position: relative;
   height: inherit;
-  min-height: 88vh;
+  min-height: 70vh;
   width: ${(props) => props.width ?? 'inherit'};
   background-color: grey;
   overflow-y: scroll;
+  overflow-x: hidden;
 `;
 
 export default Container;
