+++ frontend/src/components/organisms/WorkspaceContent/style.ts
@@ -4,6 +4,7 @@ export const Container = styled.div`
   display: flex;
   flex-direction: column;
   width: 80vw;
+  height: 95vh;
   background-color: pink;
 `;
