+++ frontend/src/components/templates/Workspace/styles.ts
